fix wrapper bug
